refactor(plugin-sdk): relocate ambient provider helpers
