+++ src/plugin-sdk/anthropic-vertex-auth-presence.ts
